fix: keeping optimize search (#2321)

+++ src/common/infra/file_list/postgres.rs
@@ -185,14 +185,19 @@ INSERT INTO file_list (org, stream, date, file, deleted, min_ts, max_ts, records
             for file in files {
                 let (stream_key, date_key, file_name) =
                     parse_file_key_columns(file).map_err(|e| Error::Message(e.to_string()))?;
-                let ret: Option<i64> = sqlx::query_scalar(
+                let ret: Option<i64> = match sqlx::query_scalar(
                     r#"SELECT id FROM file_list WHERE stream = $1 AND date = $2 AND file = $3;"#,
                 )
                 .bind(stream_key)
                 .bind(date_key)
                 .bind(file_name)
                 .fetch_one(&pool)
-                .await?;
+                .await
+                {
+                    Ok(v) => v,
+                    Err(sqlx::Error::RowNotFound) => continue,
+                    Err(e) => return Err(e.into()),
+                };
                 match ret {
                     Some(v) => ids.push(v.to_string()),
                     None => {
@@ -204,8 +209,10 @@ INSERT INTO file_list (org, stream, date, file, deleted, min_ts, max_ts, records
                 }
             }
             // delete files by ids
-            let sql = format!("DELETE FROM file_list WHERE id IN({});", ids.join(","));
-            _ = pool.execute(sql.as_str()).await?;
+            if !ids.is_empty() {
+                let sql = format!("DELETE FROM file_list WHERE id IN({});", ids.join(","));
+                _ = pool.execute(sql.as_str()).await?;
+            }
         }
         Ok(())
     }
@@ -264,14 +271,19 @@ INSERT INTO file_list (org, stream, date, file, deleted, min_ts, max_ts, records
             for file in files {
                 let (stream_key, date_key, file_name) =
                     parse_file_key_columns(file).map_err(|e| Error::Message(e.to_string()))?;
-                let ret: Option<i64> = sqlx::query_scalar(
+                let ret: Option<i64> = match sqlx::query_scalar(
                     r#"SELECT id FROM file_list_deleted WHERE stream = $1 AND date = $2 AND file = $3;"#,
                 )
                 .bind(stream_key)
                 .bind(date_key)
                 .bind(file_name)
                 .fetch_one(&pool)
-                .await?;
+                .await
+                {
+                    Ok(v) => v,
+                    Err(sqlx::Error::RowNotFound) => continue,
+                    Err(e) => return Err(e.into()),
+                };
                 match ret {
                     Some(v) => ids.push(v.to_string()),
                     None => {
@@ -283,11 +295,13 @@ INSERT INTO file_list (org, stream, date, file, deleted, min_ts, max_ts, records
                 }
             }
             // delete files by ids
-            let sql = format!(
-                "DELETE FROM file_list_deleted WHERE id IN({});",
-                ids.join(",")
-            );
-            _ = pool.execute(sql.as_str()).await?;
+            if !ids.is_empty() {
+                let sql = format!(
+                    "DELETE FROM file_list_deleted WHERE id IN({});",
+                    ids.join(",")
+                );
+                _ = pool.execute(sql.as_str()).await?;
+            }
         }
         Ok(())
     }
